+++ src/util/alertify.js
@@ -37,7 +37,7 @@ exports.alert = (text, okFnc) => {
         .show();
 }
 exports.confirm = (text, okFnc, cancelFnc) => {
-    bootstrap().confirm(text, okFnc, cancelFnc)
+    return bootstrap().confirm(text, okFnc, cancelFnc)
         .setting({
             title: TITLE,
             defaultFocus: 'ok',