+++ src/util/localstorage.js
@@ -4,6 +4,13 @@ function getKey(key) {
 
 exports.get = key => localStorage.getItem(getKey(key));
 exports.getInt = key => parseInt(localStorage.getItem(getKey(key)));
+exports.getBool = key => {
+    const item = localStorage.getItem(getKey(key));
+    if (!item) {
+        return false;
+    }
+    return item === 'true';
+}
 exports.set = (key, value) =>
     localStorage.setItem(getKey(key), value);
 exports.getKey = getKey;